pyproject.toml
[tool.poetry]
name = "telegram-bot"
version = "0.1.0"
description = ""
authors = ["Maite Puerta Beldarrain <mpuerta004@deusto.es>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
telebot = "^0.0.5"
folium = "^0.15.1"
requests = "^2.31.0"
pandas = "^2.2.0"
fastapi-utils = "^0.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
